Replace unsigned ints with plain ints
As it is generally advised for arithmetic integer expressions to use
signed ints instead of their unsigned counterparts, all occurences of
unsigned ints have been replaced. Now it is a tad easier to check for
errors. The usage of ints raises the possibility of overflows and
therefore undefined behaviour. The CMake file enables UBSanitizers as
countermeasure.

+++ costvolume.h
@@ -12,44 +12,44 @@ U narrow(T const big);
 class CostVolume
 {
 public:
-    CostVolume(const size_t height, const size_t width, const size_t d);
+    CostVolume(const int height, const int width, const int d);
     void calculate(cv::Mat const &left,
                    cv::Mat const &right,
                    cv::Mat const &left_mask = cv::Mat(),
                    cv::Mat const &right_mask = cv::Mat(),
-                   size_t const blocksize = 5);
+                   int const blocksize = 5);
 
     /**
      * @brief slice draws a slice of the cost volume as image
      * @param scanline the scanline at which to slice
      * @return the disparity slice
      */
-    cv::Mat slice(const size_t scanline) const;
+    cv::Mat slice(const int scanline) const;
 
     /**
      * @brief slice_count reports the number of possible slices
      * @return the slice count of the cost volume
      */
-    size_t slice_count() const;
+    int slice_count() const;
 
     bool is_valid() const;
 
 private:
     //! vector of scanlines, of pixels, of matching cost
-    std::vector<size_t> m_cost_volume;
+    std::vector<int> m_cost_volume;
 
     //! If a pixel is contained in this mask all disparity values must be
     //! guessed in accordance to their neighbours
     cv::Mat m_mask;
 
     //! The height of the volume
-    size_t const m_scanline_count;
+    int const m_scanline_count;
 
     //! The amount of pixels on one scanline
-    size_t const m_pixels_per_scanline;
+    int const m_pixels_per_scanline;
 
     //! The amount of displacements of one pixel
-    size_t const m_displacements_per_pixel;
+    int const m_displacements_per_pixel;
 
     /**
      * @brief to_linear maps a cost volume coordinate into linear memory
@@ -60,9 +60,7 @@ private:
      * @param d_dim disparity count
      * @return the linear index into memory
      */
-    size_t to_linear(const size_t scanline,
-                     const size_t x,
-                     const size_t d) const;
+    size_t to_linear(const int scanline, const int x, const int d) const;
     /**
      * @brief is_masked returns if a pixel location in the cost volume is masked
      * @param pixel the location in the cost volume
